adding layout overlay back in
