+++ client/main.scss
@@ -32,6 +32,7 @@ $o-grid-ie8-rules: false;
 @import "components/flexipod/main";
 
 @import "../components/fastft/main";
+@import "../components/layout-overlay/main";
 @import "../components/layout/main";
 @import "../components/feed/main";
 @import "../components/section/main";
