+++ views/front-page.html
@@ -6,7 +6,10 @@
 {{#if @root.flags.frontPageLayoutPrototype}}
 	<div id="layout-overlay-container"></div>
 	{{> header-tabs  newsTabTitle="Top Stories"}}
-	<section id="main-body" class="main-body o-grid-container n-util-clearfix">
+	<section
+		id="main-body"
+		class="main-body o-grid-container n-util-clearfix"
+		{{#if @root.flags.frontPageLayoutTool}} data-main-content="{{json content}}"{{/if}}>
 		{{{ reactRenderToString Layout content=content layout=initialLayout }}}
 	</section>
 {{else}}
